backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.6.0
anyio==4.3.0
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.9.0
cachetools==5.5.1
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
discord.py==2.5.1
emoji==2.14.1
Flask==3.1.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
idna==3.4
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
langdetect==1.0.9
lxml==4.9.3
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.60.1
opencv-python==4.11.0.86
pandas==2.1.2
pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20
python-telegram-bot==21.10
pytz==2023.3.post1
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.37
tenacity==9.0.0
tqdm==4.66.4
typing-extensions==4.12.2
urllib3==2.0.7
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
